Fix auto quoting.
2005-04-04  Morten Welinder  <terra@gnome.org>

	* gsf/gsf-output-csv.c (gsf_output_csv_write_field): Fix auto
	quoting.

+++ gsf/gsf-output-csv.c
@@ -84,11 +84,10 @@ gsf_output_csv_write_field (GsfOutputCsv *csv, const char *field, size_t len)
 	gboolean ok;
 	const char *end;
 
-
 	g_return_val_if_fail (GSF_IS_OUTPUT_CSV (csv), FALSE);
 	g_return_val_if_fail (field != NULL, FALSE);
 
-	if (len < 0) len = strlen (field);
+	if ((ssize_t)len < 0) len = strlen (field);
 	end = field + len;
 
 	if (csv->fields_on_line && csv->separator_len)
